requirements.txt
# Dev/test/deploy
websockets
asyncio
python-can
pytest
click
pytest-asyncio
